Added mixins which HoverField & AggregateField components will use
