+++ src/js/components/mixins/DraggableFields.jsx
@@ -1,0 +1,63 @@
+'use strict';
+var sg = require('../../ctrl/signals'),
+    ctrl = require('../../ctrl'),
+    LIST_LIMIT = 5;
+
+module.exports = {
+
+  handleDragStart: function(evt) {
+    this.setState({
+      bindField: this.state.fieldDef,
+      listLimit: LIST_LIMIT,
+      showFieldTransforms: false
+    });
+    evt.dataTransfer.setData('text/plain', evt.target.id);
+    evt.dataTransfer.effectAllowed = 'link';
+    sg.set(sg.MODE, 'channels');
+    ctrl.update();
+  },
+
+  handleDragOver: function(evt) {
+    if (evt.preventDefault) {
+      evt.preventDefault(); // Necessary. Allows us to drop.
+    }
+
+    return false;
+  },
+
+  // This makes use of the bubble cursor, which corresponds to the cell signal;
+  // we're using that to figure out which channel we are closest to. The
+  // SELECTED signal indicates the mark to bind the data to.
+  handleDragEnd: function(evt) {
+    var props = this.props,
+        sel = sg.get(sg.SELECTED),
+        cell = sg.get(sg.CELL),
+        bindField = this.state.bindField,
+        dropped = sel._id && cell._id;
+
+    try {
+      if (dropped) {
+        props.bindChannel(props.dsId, bindField, sel.mark.def.lyra_id, cell.key);
+      }
+    } catch (e) {
+      console.warn('Unable to bind primitive');
+      console.warn(e);
+    }
+
+    sg.set(sg.MODE, 'handles');
+    sg.set(sg.CELL, {});
+    this.setState({bindField: null});
+
+    if (!dropped) {
+      ctrl.update();
+    }
+  },
+
+  handleDrop: function(evt) {
+    if (evt.preventDefault) {
+      evt.preventDefault(); // Necessary. Allows us to drop.
+    }
+
+    return false;
+  }
+};
